requirements.txt
altair==5.0.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
newsapi-python==0.2.7
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
praw==7.7.0
prawcore==2.3.0
protobuf==4.23.2
pyarrow==12.0.0
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.31.0
rich==13.4.1
six==1.16.0
smmap==5.0.0
streamlit==1.23.1
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
update-checker==0.18.0
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
websocket-client==1.5.3
zipp==3.15.0
